+++ src/ems-esp.cpp
@@ -69,15 +69,10 @@ Ticker showerColdShotStopTimer;
 #define SHOWER_COLDSHOT_DURATION 10 // in seconds. 10 seconds for cold water before turning back hot water
 #define SHOWER_MAX_DURATION 420000  // in ms. 7 minutes, before trigger a shot of cold water
 
-#ifdef LOGICANALYZER
-#define EMSESP_DALLAS_GPIO D1
-#define EMSESP_DALLAS_PARASITE false
-#else
 // set this if using an external temperature sensor like a DS18B20
 // D5 is the default on a bbqkees board
 #define EMSESP_DALLAS_GPIO D5
 #define EMSESP_DALLAS_PARASITE false
-#endif
 
 // Set LED pin used for showing the EMS bus connection status. Solid means EMS bus working, flashing is an error
 // can be either the onboard LED on the ESP8266 (LED_BULLETIN) or external via an external pull-up LED (e.g. D1 on a bbqkees' board)
@@ -1611,7 +1606,7 @@ void MQTTCallback(unsigned int type, const char * topic, const char * message) {
 
     uint8_t hc;
     // thermostat temp changes
-    hc = _hasHCspecified(TOPIC_THERMOSTAT_CMD_TEMP, topic);
+    hc = _hasHCspecified(TOPIC_THERMOSTAT_CMD_TEMP_HA, topic);
     if (hc) {
         float f = strtof((char *)message, 0);
         ems_setThermostatTemp(f, hc);
@@ -1620,7 +1615,7 @@ void MQTTCallback(unsigned int type, const char * topic, const char * message) {
     }
 
     // thermostat mode changes
-    hc = _hasHCspecified(TOPIC_THERMOSTAT_CMD_MODE, topic);
+    hc = _hasHCspecified(TOPIC_THERMOSTAT_CMD_MODE_HA, topic);
     if (hc) {
         if (strncmp(message, "auto", 4) == 0) {
             ems_setThermostatMode(2, hc);
@@ -1642,6 +1637,29 @@ void MQTTCallback(unsigned int type, const char * topic, const char * message) {
             return;
         }
         const char * command = doc["cmd"];
+
+        // thermostat temp changes
+        hc = _hasHCspecified(TOPIC_THERMOSTAT_CMD_TEMP, command);
+        if (hc) {
+            float f = doc["data"];
+            ems_setThermostatTemp(f, hc);
+            publishValues(true); // publish back immediately
+            return;
+        }
+
+        // thermostat mode changes
+        hc = _hasHCspecified(TOPIC_THERMOSTAT_CMD_MODE, command);
+        if (hc) {
+            const char * data_cmd = doc["data"];
+            if (strncmp(data_cmd, "auto", 4) == 0) {
+                ems_setThermostatMode(2, hc);
+            } else if ((strncmp(data_cmd, "day", 4) == 0) || (strncmp(data_cmd, "manual", 6) == 0) || (strncmp(data_cmd, "heat", 4) == 0)) {
+                ems_setThermostatMode(1, hc);
+            } else if ((strncmp(data_cmd, "night", 5) == 0) || (strncmp(data_cmd, "off", 3) == 0)) {
+                ems_setThermostatMode(0, hc);
+            }
+            return;
+        }
 
         // set night temp value
         hc = _hasHCspecified(TOPIC_THERMOSTAT_CMD_NIGHTTEMP, command);